functional MakeVariable, still need refactoring

+++ src/slogo/backend/commands/control/controlcommands/Repeat.java
@@ -11,7 +11,7 @@ public class Repeat implements ControlInterface {
 
     public double execute(TurtleHistory turtleHistory, List<Object> parameters) {
         double returnValue = 0;
-        for (int i=0; i<Integer.parseInt(parameters.get(0).toString().trim()); i++) {
+        for (int i=0; i<(int) Double.parseDouble(parameters.get(0).toString().trim()); i++) {
             System.out.println("In repeat, executing block: " + parameters.get(1));
             CommandBlockManager commandBlockManager = new CommandBlockManager(parameters.get(1).toString(), turtleHistory);
             returnValue = commandBlockManager.executeInstructionBlock();
